requirements.txt
pip==26.0.1
PyYAML==6.0.3
